Update index.php
If no thumbnail exists, it now displays an image that says there is no thumbnail, rather than displaying an aurora image.
I discussed this with Thomas the end of July, 2021.

Fixed issue where thumbnails were stretched horizontally on large screens.

Removed duplicate image name code.

Left in the legacy "onmouseenter" and "onmouseleave" code, but it's not used.  Thomas can determine if it should be removed (remove $y and <script> block).