Eric Claeys' changes to allsky-website (#47)
* Add files via upload

An image to be used when no thumbnail is found for archived timelapses.  This replaces the aurora thumbnail.
I discussed this with Thomas late July, 2021.

* Update index.php

Use the image as a thumbnail if unable to create a thumbnail.

* Update index.php

Use the image as a thumbnail if unable to create a thumbnail.
Fixed issue where thumbnails were stretched horizontally on large screens.

* Update index.php

If no thumbnail exists, it now displays an image that says there is no thumbnail, rather than displaying an aurora image.
I discussed this with Thomas the end of July, 2021.

Fixed issue where thumbnails were stretched horizontally on large screens.

Removed duplicate image name code.

Left in the legacy "onmouseenter" and "onmouseleave" code, but it's not used.  Thomas can determine if it should be removed (remove $y and <script> block).

* Update config.js

Removed "-resize" because the new allsky scripts (which will be submitted soon) no longer use that suffix.
Added "showOverlayAtStartup" which displays the overlay when the page first loads, rather than requiring the user to click on the icon for it.

* Update controller.js

Uses config.showOverlayAtStartup

* Update functions.php

Checks if imagecreatefromjpeg() exists and displays an appropriate message if it doesn't, then returns false.   Returns true if it could make the thumbnail.

* Update README.md

Changed default imageName.
Added showOverlayAtStartup.